Started with search function

+++ internal/tui/home/home.go
@@ -11,6 +11,7 @@ import (
 	"github.com/SimonVillalonIT/music-golang/internal/services"
 	cmds "github.com/SimonVillalonIT/music-golang/internal/tui/commands"
 	"github.com/SimonVillalonIT/music-golang/internal/tui/constants"
+	"github.com/SimonVillalonIT/music-golang/internal/tui/download"
 	fancyList "github.com/SimonVillalonIT/music-golang/internal/tui/list"
 	"github.com/SimonVillalonIT/music-golang/internal/tui/player"
 	"github.com/SimonVillalonIT/music-golang/internal/tui/playlist"
@@ -23,6 +24,7 @@ type Model struct {
 	list        fancyList.Model
 	player      player.Model
 	playlist    playlist.Model
+	search      search.Model
 	err         error
 	jsonFile    []services.Item
 	conn        *mpvipc.Connection
@@ -38,8 +40,9 @@ func New() tea.Model {
 	list := fancyList.NewModel(items)
 	player := player.NewModel()
 	playlist := playlist.NewModel()
+	search := search.New()
 
-	return Model{list: list, jsonFile: jsonfile, isConnected: false, player: player, playlist: playlist}
+	return Model{list: list, jsonFile: jsonfile, isConnected: false, player: player, playlist: playlist, search: search}
 }
 
 func (m Model) Init() tea.Cmd {
@@ -63,7 +66,7 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.err = msg
 	case tea.KeyMsg:
 		if key.Matches(msg, constants.Keymap.Move) {
-			if m.state == constants.ListState {
+			if m.state == constants.ListState && m.playlist.IsLoaded() {
 				m.state = constants.PlaylistState
 			} else {
 				m.state = constants.ListState
@@ -71,17 +74,34 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			commands = append(commands, changeState((*uint)(&m.state)))
 		}
 		if key.Matches(msg, constants.Keymap.Quit) {
-			cmds.Kill()
-			return m, tea.Quit
+			if m.state != constants.SearchState {
+				cmds.Kill()
+				return m, tea.Quit
+			} else {
+				m.state = constants.ListState
+			}
+		}
+		if key.Matches(msg, constants.Keymap.Search) {
+			m.state = constants.SearchState
 		}
 		if key.Matches(msg, constants.Keymap.Remove) && m.state == constants.PlaylistState {
 			if m.isConnected {
-				commands = append(commands, cmds.RemoveCmd(m.conn, m.playlist.GetIndex()))
+				commands = append(commands, cmds.RemoveCmd(m.conn, m.playlist.GetPosition()))
 			}
 		}
 		if key.Matches(msg, constants.Keymap.Enter) {
 			if m.isConnected {
-				commands = append(commands, cmds.PlayCmd(m.conn, m.list.SelectedItem().(services.Item)))
+				if m.state == constants.ListState {
+					commands = append(commands, cmds.PlayCmd(m.conn, m.list.SelectedItem().(services.Item)))
+				}
+				if m.state == constants.SearchState {
+					if m.search.GetLength() > 0 {
+						current := m.search.GetCurrent()
+                        commands = append(commands, cmds.SaveCmd(&m.jsonFile, current))
+					} else {
+						commands = append(commands, cmds.SearchCmd(m.search.Answer()))
+					}
+				}
 			}
 		}
 		if key.Matches(msg, constants.Keymap.Space) {
@@ -111,19 +131,27 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		}
 	}
 
-	updatedPlayer, cmd := m.player.Update(msg)
-	m.player = updatedPlayer.(player.Model)
+	if m.state == constants.SearchState {
+		updatedSearch, searchCmd := m.search.Update(msg)
+		m.search = updatedSearch.(search.Model)
+		commands = append(commands, searchCmd)
+	} else {
+		updatedPlayer, cmd := m.player.Update(msg)
+		m.player = updatedPlayer.(player.Model)
 
-	if m.state == constants.ListState {
-		updatedList, listCmd := m.list.Update(msg)
-		m.list = updatedList.(fancyList.Model)
-		commands = append(commands, listCmd)
-	}
-	updatedPlaylist, playlistCmd := m.playlist.Update(msg)
-	m.playlist = updatedPlaylist.(playlist.Model)
+		if m.state == constants.ListState {
+			updatedList, listCmd := m.list.Update(msg)
+			m.list = updatedList.(fancyList.Model)
+			commands = append(commands, listCmd)
+		}
+		updatedPlaylist, playlistCmd := m.playlist.Update(msg)
+		m.playlist = updatedPlaylist.(playlist.Model)
 
-	if m.state == constants.PlaylistState {
-		commands = append(commands, playlistCmd)
+		if m.state == constants.PlaylistState {
+			commands = append(commands, playlistCmd)
+		}
+
+		commands = append(commands, cmd)
 	}
 
 	commands = append(commands, cmd)
@@ -139,6 +167,19 @@ func (m Model) View() string {
 	if m.conn.IsClosed() {
 		return "Loading..."
 	}
+
+	if m.state == constants.SearchState {
+		return lipgloss.Place(m.width, m.height, lipgloss.Center, lipgloss.Center, m.search.View())
+	}
+
+	if m.state == constants.ListState {
+		list = lipgloss.NewStyle().Border(lipgloss.NormalBorder()).Render(list)
+		playlist = lipgloss.NewStyle().Border(lipgloss.HiddenBorder()).Render(playlist)
+	} else {
+		playlist = lipgloss.NewStyle().Border(lipgloss.NormalBorder()).Render(playlist)
+		list = lipgloss.NewStyle().Border(lipgloss.HiddenBorder()).Render(list)
+	}
+
 	main := lipgloss.JoinHorizontal(lipgloss.Left, list, playlist)
 	return lipgloss.Place(m.width, m.height, 0, 0, lipgloss.JoinVertical(lipgloss.Left, main, m.player.View()))
 }
